feat(conf) auto-generate and inject dhparams for default intermediate cipher suite
### Summary

This work is a rework of the PR #4688. Perhaps this time we can include it,
and it is modified according to feedback received back then.

This PR changes the `cipher_suite=intermediate` to also add `ssl_dhparam`
directive in our template. For `modern` this does not do it as there are
no ciphers that need it. Also for `old` it does not do it as Mozilla's old
recommends `1024` bits params, and there are no secure predefined one found
in: RFC 7919: https://tools.ietf.org/html/rfc7919#appendix-A.

Also it does not generate it for `custom` cipher suites, but allows custom
suites to specify shorthands:

- ffdhe2048
- ffdhe3072
- ffdhe4096
- ffdhe6144
- ffdhe8192

```
KONG_NGINX_<HTTP|STREAM>_SSL_DHPARAM=<shorthand> # or
KONG_NGINX_<HTTP|STREAM>_SSL_DHPARAM=<shorthand>
```

Note the above only works with `old` and `custom` (`modern` removes this
and `intermediate` forcibly sets it to `ffdhe2048`).

### Issues Resolved

Fix #6454